fix rpmlint

+++ plugin/app/plugins/pipeline_plugin/tools/scm/git/templates/hooks/validators/spec.sh
@@ -10,7 +10,11 @@ function validate() {
   is_rpm_spec_file || return 126
   command -v rpmlint >/dev/null || return 126
   rm -f rpmlint.log
-  rpmlint "$file_name" | tee -a rpmlint.log
+  if [[ -f  "${CDN_BUILD_LIB}/linters/rpmlintrc" ]]; then
+    rpmlint -f  "${CDN_BUILD_LIB}/linters/rpmlintrc" "$file_name" | tee -a rpmlint.log
+  else
+    rpmlint "$file_name" | tee -a rpmlint.log
+  fi
   #filter by number errors. RPM_SOURCE_DIR isnot a error
   exit_code=$(<rpmlint.log grep -v "E: use-of-RPM_SOURCE_DIR$" | grep ":" | cut -d':' -f2- | grep -c -v " W:")
   rm -f rpmlint.log
